[TIR][Schedule] Scoped CacheRead/Write producing compact region (#15236)
This PR enhances CacheRead/Write so that when a cache operation is
performed under an inner block, the generated cache buffer will have
the shape as compact as possible, by region consumption analysis.

The motivation of this change comes from the needs of dynamic shape TIR
scheduling, in which case we may isolate a "static shape" internal block
using blockize, and do further scheduling inside the internal block. For
such cases, the current CacheRead/Write inside the static-shape block
will still produce dynamic-shape cache buffers, which is not ideal for
analysis and subsequent scheduling.

One thing that worths noting is that, to ensure the IR correctness after
inserting the cache block, we will only compact the cache buffer when
all the consumer blocks of the read buffer (for CacheRead) or the write
buffer (for CacheWrite) are children blocks of the cache block insertion
location. Otherwise we will insist allocating the full-size cache
buffer.

Co-authored-by: Bohan Hou <spectrometerh@gmail.com>