[TIR] Return error code from kernels in SplitHostDevice    (#15241)
* [TVMScript] Handle parsing of PrimFunc calls with non-void return

Prior to this commit, the return type of all internal function calls
was hard-coded as `"void"`.  After this commit, the `GlobalVar`
representing the internal function has type annotation based on the
callee's signature, which is then used as the return type of the
internal call.

* Update CallNode return type in MakeUnpackedAPI

* [TIR] Return error code from kernels in SplitHostDevice

Some codegen types delegate to `CodeGenCPU` for their compute kernels,
as they may delegate work to packed functions.  Because `CodeGenCPU`
assumes that it can return an error code at any point (e.g. when
launching a parallel for loop), the compute kernel should return an
error code.

* [TIR] Remove builtin::ret(0) from device-side kernel

* Restrict the int32 return type to targets that need to propagate errors

* Updated unit tests for CPU-specific checks