[Unity] Preserve shape parameters in LazyTransformParams (#15563)
Prior to this commit, any symbolic variables found in the
transform_params function that are not provided by input tensors
result in undefined symbolic variables.  This commit updates
LazyTransformParam to only delegate tensor parameters to `"get_item"`
and `"set_item"`.

This can be useful when producing sharded transformed parameters in a
multi-GPU setup, where a `rank` symbolic variable may be provided to
specify a subset of the weights to be used.