[Unity] Implement relax.Function.bind_symbolic_vars (#15509)
* [Unity] Implement relax.Function.bind_symbolic_vars

If a function has dynamic shape parameters, it can be useful to
replace them with static parameters (e.g. when producing several
models within the same family).  This commit introduces a utility
function `relax.Function.bind_symbolic_vars`, which allows symbolic
variables to be replaced with static values.

This is a related to the parameter binding done in
`relax.transform.BindParam`, but does not require the bound parameter
to be fully static data array.

* Updating ExprBinder to use tir::Substitute

Previously, `ExprBinder` only checked whether a `PrimExpr` was a
symbolic variable to be replaced, but did not handle cases where a
`PrimExpr` contained a symbolic variable to be replaced.  As a result,
when binding symbolic variables `{N: 16}`, a shape of `[N,2*N]` would be
updated to `[16,2*N]` instead of `[16,32]`.  This commit updates
`ExprBinder` to use `tir::Substitute` to ensure all occurrences of the
symbolic variable are replaced.

* Special case for updating symbolic vars in strided_slice attrs

* Added IRModule pass to bind symbolic vars

* Update unit test to include pytest

* Co-authored-by: Sunghyun Park <sunggg@umich.edu>

* Correct match mode in kProvideDefinitions context

* Clean up implementation with VisitMode as a bitflag