[Hopper TMA] Add intrinsic to create barriers for synchronization (#15684)
This PR adds an intrinsic to create barriers that can be used with existing barrier intrinsics for synchronization. The prior method of barrier allocation was to use alloc_buffer e.g. as follows barrier = T.alloc_buffer([1], "uint64", scope="shared") and then pass the pointer and offset to that barrier allocation for use in the barrier intrinsics. This was a functional interface, but also caused problems with alignment of other non-barrier shared memory allocations. See removed workarounds marked with TODO in the tests in this PR. At the expense of the additional create_barriers intrinsic we get a simplified interface using barrier ID rather than pointer / offset passed to the barrier intrinsics and more low-level codegen control which is used in this PR to solve the alignment issue.

* [Hopper TMA] Add intrinsic to create barriers for synchronization

* pad barrier alignment to avoid runtime alignment errors

* CHECK vs ICHECK; barrier alignment = 16 plus comment

* CHECK_EQ compile error