[Unity] Implemented SameShapeConstraint for dataflow pattern matches (#15694)
* [Unity] Implemented SameShapeConstraint for dataflow pattern matches

Prior to this commit, a shape could be explicitly specified using the
`ShapePattern`, but could not be specified relative to the shape of
another expression.  As a result, patterns with restricted shapes
became very difficult to

This commit implements `SameShapeConstraint`, which can be applied
between any patterns that participate in the match.  For example,
matching against `R.add(lhs,rhs)` where `lhs` and `rhs` have the same
non-broadcasted shape.  Because these constraints operate between
patterns that do not necessarily share a consumer/producer
relationship, they could not previously be expressed using the
existing `PairCons` functionality, and were instead implemented in
terms of a new `DFConstraint` base class.

* Removed empty line

* Update API based on PR discussion

Provide a `AsPrimExpr` instead of `IsConstraintSatisfied` function for
constraints.  A constraint can return a necessary-and-sufficient
condition, or a necessary-but-not-sufficient condition, and the
calling scope can decide how to interpret those results.

* Lint fix

* lint fixes