[Unity] Delegate DataflowVar visitor to Var by default (#15688)
* [Unity] Delegate DataflowVar visitor to Var by default

Prior to this commit, writing a subclass of `relax::ExprVisitor` or
`relax::ExprMutator` required separate overrides for visiting a
`relax::DataflowVar` and a `relax::Var`.  In the majority of cases,
these two types should be treated identically, and failure to handle a
`DataflowVar` equivalently would be a bug.

This commit updates the `relax::ExprVisitor` and `relax::ExprMutator`
base classes to visit `DataflowVar` by delegate to the visitor of
`relax::Var`.  As a result, any derived class that overrides the
`relax::Var` visitor will also update the behavior for
`relax::DataflowVar`.  A pass that requires different behavior for
`relax::Var` and `relax::DataflowVar` can still explicitly override
both methods in order to provide different behavior.

* Updates to Normalizer, VisitVarDef

* Removed DataflowVar visitor in VarVisitor

It was introduced in https://github.com/apache/tvm/pull/15698, before
the more general fix implemented in the current PR.