[Unity] Propagate extra symbolic vars through LiftTransformParams (#15699)
* [Unity][Analysis] Implemented DefinableTIRVarsInStructInfo

The existing utility `TIRVarsInStructInfo` returns all TIR variables,
regardless of whether they are suitable for a variable definition, or
are usage sites.  This utility walks over the struct info once,
returning both the definable symbolic variables and the used symbolic
variables.

* [Unity][Analysis] Accept relax::Expr arg in Defined/FreeSymbolicVars

Prior to this commit, this utility could only be used with a
`relax::Function` argument.  This allows individual expressions to be
inspected, even if they are not part of a complete function.

* [Unity] Propagate symbolic variables in LiftTransformParams

* Updated LiftTransformParams to use support::OrderedSet

* Fixed import after rebase