[Community] Shuai Yuan -> Committers (#16162)
[Community] Shuai Yuan -> committers

Please join me in welcoming Shuai Yuan(@ysh329) as a new committer in TVM.
His persistent assistance in the release process is noteworthy.
Specifically, his involvement in drafting release notes and coordinating
the release schedules for the recent three releases (v0.12, v0.13, and
v0.14) is commendable. Although Shuai does not currently hold the
position of a reviewer, his contributions significantly surpass the
requirements for a committer role. It is crucial to recognize and
appreciate contributions that extend beyond mere code production.

- [Github Issues](https://github.com/apache/tvm/issues?page=1&q=is%3Aissue+author%3Aysh329)
- [Commits History](https://github.com/apache/tvm/pulls?q=is%3Apr+author%3Aysh329)
- [Code Review](https://github.com/apache/tvm/pulls?utf8=%E2%9C%93&q=reviewed-by:ysh329)
- [Community Forum Summary](https://discuss.tvm.apache.org/u/ysh329/summary)