[Unity][Bugfix] Handle symbolic matching with non-structural match (#15994)
* [Unity][Bugfix] Handle symbolic matching with non-structural match

Prior to this commit, `FuseTIR` required the symbolic shape of an
argument to be a structural match to the symbolic shape of a
parameter.  As a result, erroneous failures to match would occur when
the dynamic shape parameter contains expressions, and the static shape
argument does not.  For example, a parameter `[n, m, n*m]` is not
structurally matched to an argument `[16, 16, 256]`, but is a valid
substitution.  This commit updates `FuseTIR` to handle these cases.

* Avoid repeated initialization of arith::Analyzer