[Unity][Transform] Implement Relax function inlining (#16194)
* [Unity][Transform] Implement Relax function inlining

This commit adds the ability to inline one Relax function into
another.  This is provided with two APIs, one which allows explicit
specification of the functions to be inlined, and one which inlines
every private Relax function in an `IRModule`.  Function inlining with
explicit replacements is intended for use by an end-user when building
a model function (e.g. inlining of a rotary embedding), and the second
is intended for use as part of a generic optimization pipeline.

* lint fix

* Use DetectRecursion from relax/analysis.h

* Added test case for error on mutually-recursive functions