[Unity][MSC] Avoid depending on trivial bindings in Relax intermediate (#16349)
* [Unity][MSC] Avoid depending on trivial bindings in Relax intermediate

The conversion from tensorflow to MSC is done by first converting from
tensorflow to relay, then converting from relay to executable python
code, executing that python code to generate relax, and finally
converting from relax to MSC.  During the relax phase of this
conversion, some relax `IRModule` are applied, including
`FuseOpsByPattern`.

The test cases in `test_msc/test_translate_tensorflow.py` rely on
`FuseOpsByPattern` preserving trivial bindings (e.g. `var_1 = var_2`)
in the relax IRModule.  If these trivial bindings are removed by
`CanonicalizeBindings`, then the test cases in this file fail.  The
presence or absence of trivial bindings `FuseOpsByPattern` should be
considered an implementation detail, and relax passes should not be
required to preserve trivial bindings.

This PR updates the relay to executable python step of the tensorflow
to MSC conversion, to remove trivial bindings and output a variable
name that matches the expected value in the test case.  While not an
ideal resolution, as other variable name changes could still
reintroduce the same test failures, it ensures that `FuseOpsByPattern`
may canonicalize bindings as an internal pre- or post-processing step
without breaking these unit tests.

* Update implementation to remove dataflow block in MSC codegen

The potential for duplicate variable names was introduced by having
the `block_builder.emit_output` call, which is only required to export
values from a dataflow block.  The dataflow block is not used in any
later MSC conversion, and its removal avoids this re-export of
variables.

If the dataflow block is required in the future, it can be generated
using `tvm.relax.transform.ConvertToDataflowBlock`.

* Make failing test cases be close to the same structural form

* Updated tests to validate output after compilation

* Lint fixes