[Debug] Improve error message for codegen pattern mismatches (#16587)
If there is a mis-match between the pattern used in `FuseOpsByPattern`
and the expected relax function used in a relax codegen
function (e.g. `"relax.ext.cublas"`), this will raise an error when
attempting to unpack the primitive relax function.  This commit
improves the error message provided when this occurs, stating what
pattern was expected, and what function was expected to match the
pattern.