[DeviceAPI] Support "GetCurrentStream" (#16689)
This PR introduces a new function `GetCurrentStream`to device API,
which returns the current stream of the given device.

Meanwhile, this PR updates the "CreateStream" of CUDA to creating
a non-blocking stream, so that the execution on this stream can
overlap with the execution of other streams.

This PR also changes the `GPUCopy` of CUDA device API to always
using `cudaMemcpyAsync`.