[SLM] Allow modules to define pre-processing of weights (#16757)
* [SLM] Allow TensorStructInfo to specify parameter in export

Prior to this commit, the parameter specification for SLM tensor
needed to be passed as a `nn.spec.Tensor`.  As this object is only
used to construct a `relax.TensorStructInfo`, and has the same fields
as a `relax.TensorStructInfo`, this commit allows the parameter
specification to be passed as a `relax.TensorStructInfo`.

* Resolve breakage in unit tests

* [SLM] Use `CopyWithNewVars` to de-duplicate symbolic variables

Prior to this commit, a `nn.spec.Tensor`'s shape had special handling
to ensure that symbolic variable were not reused across multiple
functions.  This commit updates this to instead be performed using the
`CopyWithNewVars` function.

* [SLM] Allow modules to define pre-processing of weights

Prior to this commit, the weights used by `nn.Module` instances were
required to be `nn.Parameter` instances.  This commit allows the
weights to instead be `nn.Tensor` instances, defined in terms of other
`nn.Parameter` weights.  This allows a model to define both the
original weights that would be present in an external
checkpoint (e.g. a Pytorch or Safetensors file), and the
pre-processing that should be performed on those weights.

* Undo portions that would introduce R.Tensor to nn.Module

* Remove unit tests that were related to TensorStructInfo