[Relax] Allow DeadCodeElimination within ApplyPassToFunction (#16801)
The `tvm.ir.transform.ApplyPassToFunction` allows a transform to be
applied selectively to some portions of a `IRModule`, without applying
to the entire `IRModule`.  For example, to apply an optimization
pass (e.g. `relax.transform.ExpandMatmulOfSum`) or an
interface-altering pass (e.g. `relax.transform.BundleModelParams`) to
specific functions.  It does so by generating an intermediate
`IRModule` containing only the functions specified, applying the
transform to that intermediate, then merging the results.

When using `ApplyPassToFunction` to apply `DeadCodeElimination`, or a
pipeline containing `DeadCodeElimination`, this intermediate
`IRModule` may contain calls to `GlobalVar` instances that are not
within the intermediate `IRModule`.  Prior to this commit, this
resulted in an error being thrown when collecting the call graph.
This commit updates `DeadCodeElimination` to instead handle incomplete
call-graph collection.