[KVCache] Initialize one extra page than specified (#16849)
This PR udpates PagedKVCache to initialize one more page than
specified via constructor. The reason is that applications usually
depends the number of free pages (returned from `GetNumAvailablePages`)
to decide the KV cache operation policy. If there is no this extra
page, the KV cache will tell "no available" pages even when the
last allocated pages are not full, which may give the applications
an illusion that the KV cache is already completely full, and cause
further issues.