[Relax] Handle binary operations between Tensor and PrimValue (#16827)
* [Relax] Handle binary operations between Tensor and PrimValue

Prior to this commit, binary operations were only defined between two
tensors.  This commit allows binary operations to apply between a
tensor and a `relax::PrimValue`.

When inferring the output `StructInfo`, binary operations with a
`PrimValue` produce the same output as using a 0-d tensor.  When
legalizing operations containing a `PrimValue`, they are lowered to
primitive TIR arguments.

* Fix unit tests

* Restore ICHECK for scalar TIR variable

* Fix a few more unit tests

* Remove handling of ObjectStructInfo

* Undo commenting-out of test cases

* Update for improved error messages

* Fix failing unit tests

* Fix unit test