[Fix][Dlight] Fix GeneralReduction for log-sum-exp (#16923)
This PR fixes the GeneralReduction dlight rule so that it can support
scheduling log-sum-exp function.

Prior to this issue, the rule makes a strong assumption on the pattern
of the given function, which allows scheduling softmax, but fails to
schedule log-sum-exp due to pattern mismatch. This PR enhances the rule
and makes it able to match the pattern of log-sum-exp and apply
subsequent scheduling.

A regression test is added.