[CI] Use LLVM17 for tests on `ci_cpu` (#16931)
Changes the config script to build TVM with LLVM17. This enables tests
for #16921.

There was a failing codegen test when updating to LLVM 17, it seems it
stopped producing vectorized code with LLVM 16. I have checked the
same test with LLVM 18 and it now correctly produces vectorized code. I
made an attempt to track down the commit that fixed the issue in LLVM
but didn't have any success. Therefore, I think the best solution is
to skip the test until a more recent version of LLVM is used in CI.