[Bugfix] Update FAttrsGetter to return Map<String, ObjectRef> (#17096)
Prior to this commit, `FAttrsGetter` was defined as a function that
returned `Map<String, String>`.  However, it is used to define
attributes in a `Map<String, ObjectRef>`, and in some cases is used to
define attributes whose value is a dictionary (e.g. `msc_attrs_getter`
in `python/tvm/contrib/msc/core/transform/pattern.py`).

This commit updates the type signature of `FAttrsGetter` to match its
usage, returning a `Map<String, ObjectRef>`.